backend/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "pfa-backend"
version = "0.1.0"
description = "Personal Financial Analyst backend (CSV/PDF ingestion + ledger)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "psycopg[binary]>=3.2",
  "python-multipart>=0.0.9",
  "pypdf>=5",
]

[project.optional-dependencies]
dev = ["pytest>=8", "httpx>=0.27", "langsmith>=0.3,<1", "fpdf2>=2.8"]

[tool.setuptools.packages.find]
where = ["."]
include = ["pfa*"]

[tool.setuptools.package-data]
pfa = ["schema.sql"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "integration: requires DATABASE_URL and Postgres",
]
